requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
langchain==0.1.14
openai==1.23.2
python-dotenv==1.0.1
google-generativeai==0.3.2
